feat(library): Library Renaming
Update library path
